Split crate into "lib" and "bin"
